% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_false_negative_rate}
\alias{compute_false_negative_rate}
\title{Compute the false negative rate}
\usage{
compute_false_negative_rate(IRR, proportion_selected)
}
\arguments{
\item{IRR}{The inter-rater reliability}

\item{proportion_selected}{The proportion of selected candidates}
}
\value{
The false negative rate
}
\description{
Compute the false negative rate
based on the inter-rater reliability and the
proportion of selected candidates
}
