% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_proportion_of_correctly_selected}
\alias{compute_proportion_of_correctly_selected}
\title{Compute the proportion of correctly selected candidates}
\usage{
compute_proportion_of_correctly_selected(IRR, proportion_selected)
}
\arguments{
\item{IRR}{The inter-rater reliability}

\item{proportion_selected}{The proportion of selected candidates}
}
\value{
The proportion of correctly selected candidates
}
\description{
Compute proportion of correctly selected candidates
based on the inter-rater reliability and the
proportion of selected candidates
}
