% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_IRR2FPR.R
\name{IRR2FPR_internal}
\alias{IRR2FPR_internal}
\alias{IRR2FPR_ui}
\alias{IRR2FPR_server}
\title{\code{IRR2FPR} module (internal documentation)}
\usage{
IRR2FPR_ui(id, imports = NULL, ...)

IRR2FPR_server(id, imports = NULL, ...)
}
\arguments{
\item{id}{\emph{character}, the ID assigned by ShinyItemAnalysis. \strong{Do not set any
default value!}}

\item{imports}{\emph{list}, reactive objects exported by ShinyItemAnalysis. See
\code{vignette("imports", "SIAtools")} for more details on how to use objects
from the ShinyItemAnalysis app.}

\item{...}{additional parameters (not used at the moment).}
}
\description{
This is the internal documentation of your module that is not included in the
help index of the package. You may include your notes here. For a
\link[=IRR2FPR]{user-facing help page}, please edit the documentation
above.

Note that even being internal, a curious user can still discover this
internal help page. To prevent that, include the \verb{@noRd} \code{{roxygen2}} tag
below the line with \verb{@keywords} tag.

If your module uses any external packages, such as ggplot2,
\strong{you have to declare the imports} with the \verb{@importFrom} tag and include
the package in the DESCRIPTION. See
\url{https://r-pkgs.org/dependencies-in-practice.html} for more details.

You can preview your module using
\code{SIAtools::preview_module("IRR2FPR")}.

See \code{vignette("developing_modules", "SIAtools")} vignette for further details.
}
\keyword{internal}
