% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIAModuleIRR2FPR-package.R
\docType{package}
\name{SIAModuleIRR2FPR-package}
\alias{SIAModuleIRR2FPR}
\alias{SIAModuleIRR2FPR-package}
\title{SIAModuleIRR2FPR: Shiny Item Analysis Module for computing False Positive Rate from Inter-Rater Reliability}
\description{
Implements a Shiny Item Analysis module for computing False Positive Rate from Inter-Rater Reliability based on Barto코 & Martinkov치 (2022) \doi{10.48550/arXiv.2207.09101}.
}
\author{
\strong{Maintainer}: Franti코ek Barto코 \email{f.bartos96@gmail.com} (\href{https://orcid.org/0000-0002-0018-5573}{ORCID})

}
\keyword{internal}
