% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spearman_brown_formula}
\alias{spearman_brown_formula}
\title{Compute IRR from the Spearman-Brown formula}
\usage{
spearman_brown_formula(IRR_1, n_raters)
}
\arguments{
\item{IRR_1}{The inter-rater reliability of the first rater}

\item{n_raters}{The number of raters}
}
\value{
The inter-rater reliability
}
\description{
Compute the inter-rater reliability
based on the Spearman-Brown formula
}
