% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_IRR2FPR.R
\name{IRR2FPR}
\alias{IRR2FPR}
\title{Interactive Module for Inter-Rater Reliability to False Positive Rate Conversion}
\description{
This module allows users to convert inter-rater reliability (IRR) to false positive rate (FPR)
as described in Bartoš and Martinková (2022).
}
\references{
Bartoš, F., & Martinková, P. (2022). Selecting applicants based on multiple ratings:
Using binary classification framework as an alternative to inter-rater reliability. \doi{10.48550/arXiv.2207.09101}
}
\author{
František Bartoš
}
\concept{SIAmodules}
